How to prints the first 50 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	int cnt1, strd1, strd2, fbnc;
+	long int cnt1, strd1, strd2, fbnc;
 
 	for (cnt1 = 1; cnt1 <= 50; cnt1++)
 	{
